Cleanup bitwise and arithmetic operations
