+++ src/function/scalar/operators/CMakeLists.txt
@@ -1,15 +1,7 @@
 add_library_unity(duckdb_func_ops
                   OBJECT
-                  add.cpp
-                  subtract.cpp
-                  multiply.cpp
-                  divide.cpp
-                  mod.cpp
-                  left_shift.cpp
-                  right_shift.cpp
-                  bitwise_and.cpp
-                  bitwise_or.cpp
-                  bitwise_xor.cpp)
+                  arithmetic.cpp
+                  bitwise.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_func_ops>
     PARENT_SCOPE)
