--- src/function/scalar/operators/bitwise_or.cpp
@@ -1,19 +1,0 @@
-#include "duckdb/function/scalar/operators.hpp"
-#include "duckdb/common/vector_operations/vector_operations.hpp"
-
-using namespace duckdb;
-using namespace std;
-
-static void bitwise_or_function(ExpressionExecutor &exec, Vector inputs[], index_t input_count,
-                                BoundFunctionExpression &expr, Vector &result) {
-	result.Initialize(inputs[0].type);
-	VectorOperations::BitwiseOR(inputs[0], inputs[1], result);
-}
-
-void BitwiseOrFun::RegisterFunction(BuiltinFunctions &set) {
-	ScalarFunctionSet functions("|");
-	for (auto &type : SQLType::INTEGRAL) {
-		functions.AddFunction(ScalarFunction({type, type}, type, bitwise_or_function));
-	}
-	set.AddFunction(functions);
-}
