--- src/function/scalar/operators/mod.cpp
@@ -1,21 +1,0 @@
-#include "duckdb/function/scalar/operators.hpp"
-#include "duckdb/common/vector_operations/vector_operations.hpp"
-
-using namespace duckdb;
-using namespace std;
-
-static void mod_function(ExpressionExecutor &exec, Vector inputs[], index_t input_count, BoundFunctionExpression &expr,
-                         Vector &result) {
-	result.Initialize(inputs[0].type);
-	VectorOperations::Modulo(inputs[0], inputs[1], result);
-}
-
-void ModFun::RegisterFunction(BuiltinFunctions &set) {
-	ScalarFunctionSet functions("%");
-	for (auto &type : SQLType::NUMERIC) {
-		functions.AddFunction(ScalarFunction({type, type}, type, mod_function));
-	}
-	set.AddFunction(functions);
-	functions.name = "mod";
-	set.AddFunction(functions);
-}
